+++ test/test.js
@@ -9,6 +9,7 @@ const api = new MSFS_API();
 (async function tryConnect() {
   console.log(`Testing call prevention prior to connection`);
   await testAPriori();
+  console.log(`Awaiting connection`);
   api.connect({
     autoReconnect: true,
     retries: Infinity,
@@ -67,23 +68,26 @@ async function connect(handle) {
   console.log(`${ALL_AIRPORTS.length} total airports on the planet`);
 
   const { NEARBY_AIRPORTS } = await api.get(`NEARBY_AIRPORTS`);
-  console.log(`${NEARBY_AIRPORTS.length} nearby airports`);
+  console.log(`${NEARBY_AIRPORTS.length} airports in local reality bubble`);
 
+  const radius = 10;
+  const near = await api.get(`NEARBY_AIRPORTS:${radius}`);
+  const nearList = near[`NEARBY_AIRPORTS:${radius}`];
+  console.log(
+    `${nearList.length} airports found within a ${radius}NM radius around the plane`
+  );
+
+  console.log(`getting ${NEARBY_AIRPORTS[0].icao}`);
   const airportData = await api.get(`AIRPORT:${NEARBY_AIRPORTS[0].icao}`);
-  console.log(JSON.stringify(airportData, null, 2));
-  console.log(JSON.stringify(await api.get(`AIRPORT:CYYJ`), null, 2));
+  console.log(`result:`, JSON.stringify(airportData, null, 2));
 
-  return;
+  api.on(SystemEvents.AIRPORTS_IN_RANGE, (data) => {
+    console.log(`New in range:`, data);
+  });
 
-  // const inRange = api.on(SystemEvents.AIRPORTS_IN_RANGE, (data) => {
-  //   console.log(data);
-  //   inRange();
-  // });
-
-  // const outOfRange = api.on(SystemEvents.AIRPORTS_OUT_OF_RANGE, (data) => {
-  //   console.log(data);
-  //   outOfRange();
-  // });
+  api.on(SystemEvents.AIRPORTS_OUT_OF_RANGE, (data) => {
+    console.log(`Out of range:`, data);
+  });
 
   try {
     console.log(`Quick "unknown var" test`);